+++ Assets/_GameAssets/Scripts/Collectibles/GoldWheatCollectible.cs
@@ -5,12 +5,11 @@ public class GoldWheatCollectible : MonoBehaviour,ICollectible
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     [SerializeField] private PlayerController _playerController;
 
-    [SerializeField] private float _duration = 10;
-    [SerializeField] private float _movementSpeedBuff = 10;
+    [SerializeField] private WheatDesignSO _goldWheatDesignSO; 
 
     public void Collect()
     {
-        _playerController.ApplyGoldWheatEffects(_movementSpeedBuff, _duration);
+        _playerController.ApplyGoldWheatEffects(_goldWheatDesignSO.IncreaseDecreaseAmount, _goldWheatDesignSO.Duration);
         Destroy(gameObject);
     }
 }